+++ nucleus/validate/eval_functions/available_eval_functions.py
@@ -342,7 +342,7 @@ class SegmentationToPolyIOUConfig(EvalFunctionConfig):
 class SegmentationToPolyMAPConfig(EvalFunctionConfig):
     def __call__(
         self,
-        iou_threshold: float = 0.5,
+        iou_thresholds: Union[List[float], str] = "coco",
         annotation_filters: Optional[
             Union[ListOfOrAndFilters, ListOfAndFilters]
         ] = None,
@@ -387,7 +387,7 @@ class SegmentationToPolyMAPConfig(EvalFunctionConfig):
                 Finally, the most outer list combines these filters as a disjunction (OR).
         """
         return super().__call__(
-            iou_threshold=iou_threshold,
+            iou_thresholds=iou_thresholds,
             annotation_filters=annotation_filters,
             prediction_filters=prediction_filters,
             **kwargs,